@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide NestedScrollView;
 import 'package:flutter/scheduler.dart';
-import 'package:extended_nested_scroll_view/extended_nested_scroll_view.dart' as extended;
+import 'package:extended_nested_scroll_view/extended_nested_scroll_view.dart';
 import 'package:timer/noscale_flexible_spacebar.dart';
 import 'package:timer/timer_painter.dart';
 
@@ -12,36 +12,31 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
-
   final Color _primaryColor = Colors.deepOrangeAccent; //主色
-  Animation<int> _intAnim;  //控制圈圈
-  AnimationController _animationController; //控制圈圈
-  bool _isActivated = false;  //是否点过了开始
-  bool _isAnimating = false;  //理解为是否在转圈圈
-  bool _isAppbarExpanded = true;  //bar是否为展开状态
-  final int _lineNum = 180;    //控制圈圈的竖线的数量
+  late Animation<int> _intAnim =
+      IntTween(begin: 0, end: _lineNum).animate(_animationController);
+
+  late AnimationController _animationController = AnimationController(
+      duration: const Duration(milliseconds: 2000), vsync: this); //控制圈圈
+  bool _isActivated = false; //是否点过了开始
+  bool _isAnimating = false; //理解为是否在转圈圈
+  bool _isAppbarExpanded = true; //bar是否为展开状态
+  final int _lineNum = 180; //控制圈圈的竖线的数量
   final double _appbarExpandedHeight = 280; //bar展开的高度
   final double _appbarHeight = 120; //bar折叠后的高度
-  double _fontSize = 38;  //计时不够一小时是不显示 小时(hour) 的，当显示 小时 时将这个变量调小以适应界面
-  final List<Duration> _durationList = <Duration>[];  ///存放每次点击计次按钮时的[_currentDuration]
-  ValueNotifier<Duration> _currentDuration = ValueNotifier(Duration.zero);  //时刻增长着的，已经累计的时间
-  DateTime _startTime;  //按下开始时
-  DateTime _pauseTime;  //按下暂停时
-  DateTime _continueTime;  //按下继续时
-  final ScrollController _sc = ScrollController();
-  Ticker _ticker; //在此ticker的回调中触发计时功能
-  double get _pinnedHeaderHeight => _isActivated ? 120 : 290;
-  var _runAnimateTo;  //sc在执行animateTo()的过程中被滚动打断滚动功能就崩溃了，这个变量用来解决（凑合）这个问题
-  @override
-  void initState() {
-    super.initState();
-    _ticker = createTicker(_onTick);  //构造一个Ticker并添加了回调
-    //设置了圈圈转一圈的时间 2000 毫秒
-    _animationController = AnimationController(duration: const Duration(milliseconds: 2000), vsync: this);
-    ///实际上控制了圈圈的旋转效果[_intAnim.value]是最长彩色短线的位置，变换后就形成了动画效果
-    _intAnim = IntTween(begin: 0, end: _lineNum).animate(_animationController);
+  double _fontSize = 38; //计时不够一小时是不显示 小时(hour) 的，当显示 小时 时将这个变量调小以适应界面
+  final List<Duration> _durationList = <Duration>[];
 
-  }
+  ///存放每次点击计次按钮时的[_currentDuration]
+  ValueNotifier<Duration> _currentDuration =
+      ValueNotifier(Duration.zero); //时刻增长着的，已经累计的时间
+  late DateTime _startTime; //按下开始时
+  late DateTime _pauseTime; //按下暂停时
+  late DateTime _continueTime; //按下继续时
+  final ScrollController _sc = ScrollController();
+  late Ticker _ticker = createTicker(_onTick); //在此ticker的回调中触发计时功能
+  double get _pinnedHeaderHeight => _isActivated ? 120 : 290;
+  var _runAnimateTo; //sc在执行animateTo()的过程中被滚动打断滚动功能就崩溃了，这个变量用来解决（凑合）这个问题
 
   @override
   void dispose() {
@@ -70,11 +65,12 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
       ),
     );
   }
+
   //旋转计时器以及计次列表
   Widget _clockAndDurationList() {
     return NotificationListener<ScrollNotification>(
       onNotification: _handleScrollNotification,
-      child: extended.NestedScrollView(
+      child: ExtendedNestedScrollView(
           controller: _sc,
           headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
             return <Widget>[
@@ -83,7 +79,7 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
                 toolbarHeight: _appbarHeight,
                 elevation: 0,
                 backgroundColor: Theme.of(context).scaffoldBackgroundColor,
-                flexibleSpace: ValueListenableBuilder(
+                flexibleSpace: ValueListenableBuilder<Duration>(
                     valueListenable: _currentDuration,
                     child: Column(
                       children: [
@@ -91,32 +87,44 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
                           width: _appbarExpandedHeight,
                           height: _appbarExpandedHeight,
                           child: CustomPaint(
-                            painter: TimerPainter(_lineNum, _intAnim, _isActivated),
+                            painter:
+                                TimerPainter(_lineNum, _intAnim, _isActivated),
                             child: Column(
                               mainAxisAlignment: MainAxisAlignment.center,
                               children: [
-                                const Text('秒', style: TextStyle(color: Colors.transparent)),
+                                const Text(
+                                  '秒',
+                                  style: TextStyle(color: Colors.transparent),
+                                ),
                                 Text(
                                   ' ',
-                                  style: TextStyle(color: Colors.black, fontSize: _fontSize),
+                                  style: TextStyle(
+                                      color: Colors.black, fontSize: _fontSize),
                                 ),
-                                const Text('秒表', style: TextStyle(color: Colors.black54)),
+                                const Text(
+                                  '秒表',
+                                  style: TextStyle(color: Colors.black54),
+                                ),
                               ],
                             ),
                           ),
                         ),
                       ],
                     ),
-                    builder: (BuildContext context, value, Widget child) {
+                    builder:
+                        (BuildContext context, Duration value, Widget? child) {
                       return NoScaleFlexibleSpaceBar(
-                          centerTitle: true,
-                          titlePadding: EdgeInsets.zero,
-                          title: Center(
-                            child: DefaultTextStyle(
-                                style: TextStyle(color: Colors.black, fontSize: _fontSize),
-                                child: _dynamicDuration(value)),
+                        centerTitle: true,
+                        titlePadding: EdgeInsets.zero,
+                        title: Center(
+                          child: DefaultTextStyle(
+                            style: TextStyle(
+                                color: Colors.black, fontSize: _fontSize),
+                            child: _dynamicDuration(value),
                           ),
-                          background: child);
+                        ),
+                        background: child ?? SizedBox.shrink(),
+                      );
                     }),
               ),
             ];
@@ -127,13 +135,16 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
           body: _durationsListView()),
     );
   }
+
   //计次列表
   Widget _durationsListView() {
     return CupertinoScrollbar(
       child: DefaultTextStyle(
         style: TextStyle(color: Colors.black, fontSize: 16),
         child: ListView.builder(
-          physics: _isActivated ? AlwaysScrollableScrollPhysics() : NeverScrollableScrollPhysics(),
+          physics: _isActivated
+              ? AlwaysScrollableScrollPhysics()
+              : NeverScrollableScrollPhysics(),
           itemCount: _durationList.length,
           itemBuilder: (BuildContext context, int index) {
             return _durationItem(index);
@@ -142,28 +153,34 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
       ),
     );
   }
+
   //计次列表的每一行
   Widget _durationItem(int index) {
-    String realIndex = (_durationList.length - index).toStringAsFixedFromStart(2);
+    String realIndex =
+        (_durationList.length - index).toStringAsFixedFromStart(2);
     if (index == 0) {
       return Padding(
         padding: const EdgeInsets.symmetric(horizontal: 20),
         child: Row(
           children: [
-            Text(realIndex, style: TextStyle(color: _primaryColor, fontWeight: FontWeight.w500)),
+            Text(realIndex,
+                style: TextStyle(
+                    color: _primaryColor, fontWeight: FontWeight.w500)),
             SizedBox(height: 40, width: 16),
             ValueListenableBuilder(
               valueListenable: _currentDuration,
-              builder: (BuildContext context, Duration value, Widget child) {
+              builder: (BuildContext context, Duration value, Widget? child) {
                 return _dynamicDuration(value);
               },
             ),
             Spacer(),
             ValueListenableBuilder(
               valueListenable: _currentDuration,
-              builder: (BuildContext context, Duration currentDuration, Widget child) {
-                return Text('+${(currentDuration - _durationList.first).toStringAsMyFormat()}',
-                    style: TextStyle(fontSize: 14));
+              builder: (BuildContext context, Duration currentDuration, _) {
+                return Text(
+                  '+${(currentDuration - _durationList.first).toStringAsMyFormat()}',
+                  style: TextStyle(fontSize: 14),
+                );
               },
             )
           ],
@@ -178,13 +195,15 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
             const SizedBox(height: 40, width: 16),
             Text(_durationList[index - 1].toStringAsMyFormat()),
             const Spacer(),
-            Text('+${(_durationList[index - 1] - _durationList[index]).toStringAsMyFormat()}',
+            Text(
+                '+${(_durationList[index - 1] - _durationList[index]).toStringAsMyFormat()}',
                 style: TextStyle(fontSize: 14))
           ],
         ),
       );
     }
   }
+
   ///监听[_currentDuration]刷新[Text]的内容构成动态计时的样式
   Widget _dynamicDuration(Duration value) {
     List<String> currentDuration = value.toStringAsMyFormat().split('.');
@@ -195,12 +214,15 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
         Text.rich(
           TextSpan(text: currentDuration[0], children: <InlineSpan>[
             const TextSpan(text: '.'),
-            TextSpan(text: currentDuration[1], style: TextStyle(color: _primaryColor))
+            TextSpan(
+                text: currentDuration[1],
+                style: TextStyle(color: _primaryColor))
           ]),
         ),
       ],
     );
   }
+
   //底部的按钮区域
   Widget _bottomButton() {
     return Container(
@@ -216,7 +238,9 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
                         height: 60,
                         child: FlatButton(
                           child: Text(_isAnimating ? '计次' : '复位'),
-                          shape: CircleBorder(side: BorderSide(color: Colors.grey.shade300, width: 2)),
+                          shape: CircleBorder(
+                              side: BorderSide(
+                                  color: Colors.grey.shade300, width: 2)),
                           color: Colors.transparent,
                           onPressed: _onCountOrResetPressed,
                         ),
@@ -226,10 +250,14 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
                         height: 60,
                         child: FlatButton(
                           child: Text(_isAnimating ? '暂停' : '继续'),
-                          textColor: _isAnimating ? _primaryColor : Colors.white,
+                          textColor:
+                              _isAnimating ? _primaryColor : Colors.white,
                           shape: CircleBorder(
-                              side: BorderSide(color: Colors.grey.shade300, width: _isAnimating ? 2 : 0)),
-                          color: _isAnimating ? Colors.transparent : _primaryColor,
+                              side: BorderSide(
+                                  color: Colors.grey.shade300,
+                                  width: _isAnimating ? 2 : 0)),
+                          color:
+                              _isAnimating ? Colors.transparent : _primaryColor,
                           onPressed: _onPauseOrContinuePressed,
                         ),
                       ),
@@ -249,17 +277,20 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
           ],
         ));
   }
+
   ///通过监听滚动区域滚动状态来实现'自动'展开或收起[SliverAppBar]
   bool _handleScrollNotification(ScrollNotification notification) {
     if (notification is ScrollEndNotification) {
-      if (_sc.offset > 0 && _sc.offset < (_appbarExpandedHeight - _appbarHeight)) {
-        double target = _isAppbarExpanded ? _appbarExpandedHeight - _appbarHeight : 0;
-        Future.delayed(Duration.zero,() async{
-          if(_runAnimateTo != null){
+      if (_sc.offset > 0 &&
+          _sc.offset < (_appbarExpandedHeight - _appbarHeight)) {
+        double target =
+            _isAppbarExpanded ? _appbarExpandedHeight - _appbarHeight : 0;
+        Future.delayed(Duration.zero, () async {
+          if (_runAnimateTo != null) {
             await _runAnimateTo;
           }
-          _runAnimateTo = _sc.animateTo(
-              target, duration: const Duration(milliseconds: 80), curve: Curves.linear);
+          _runAnimateTo = _sc.animateTo(target,
+              duration: const Duration(milliseconds: 80), curve: Curves.linear);
         }).then((value) {
           if (_sc.offset == 0) {
             _isAppbarExpanded = true;
@@ -272,6 +303,7 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
     }
     return false;
   }
+
   ///[_ticker]将会触发的回调
   void _onTick(Duration elapsed) {
     DateTime now = DateTime.now();
@@ -283,6 +315,7 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
       });
     }
   }
+
   //开始 按钮
   void _onStartPressed() {
     _isActivated = true;
@@ -293,6 +326,7 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
     setState(() {});
     _animationController.repeat();
   }
+
   //暂停/继续 按钮
   void _onPauseOrContinuePressed() {
     if (_isAnimating) {
@@ -309,6 +343,7 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
     }
     setState(() {});
   }
+
   //计次/重置 按钮
   void _onCountOrResetPressed() {
     if (_isAnimating) {
@@ -326,12 +361,14 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
       setState(() {});
     }
   }
+
   //计次
   void _addACount() {
     _durationList.insert(0, _currentDuration.value);
   }
 
-  double calculateTextWidth(String value, double fontSize, double maxWidth, int maxLines) {
+  double calculateTextWidth(
+      String value, double fontSize, double maxWidth, int maxLines) {
     TextPainter painter = TextPainter(
         maxLines: maxLines,
         textDirection: TextDirection.ltr,
@@ -341,6 +378,7 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
               fontSize: fontSize,
             )));
     painter.layout(maxWidth: maxWidth);
+
     ///文字的宽度:painter.width
     return painter.width;
   }
